+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Reverted to allowing single character emoji shortcodes.
 - Don't pollute PostgreSQL log with `post_reaction` table constraint violation errors.
 - Don't import post when `reject-data` rule for target domain exists.
+- Drop delivery if `reject-data` rule for target domain exists.
 
 ### Fixed